# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "金丝雀蜜罐 CanaryGuard AntiFraud",
    "description": "蜜罐反诈双形态主控：REST API 同一内核，WebUI/MCP 均为客户端（D1）。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/accounts/check": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Account Check",
        "description": "账号风险速查：绿黄红 + 证据链 + 共现账号。\n\nR4-F3：匿名调用（无有效 admin key）→ 只读评估不落库（persist=False，\n防匿名灌入账号库；已有档案仍可读）；已认证调用 → 维持画像 upsert。",
        "operationId": "account_check_api_v1_accounts_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Account Check Api V1 Accounts Check Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{url_name}": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Get Account",
        "description": "查询已建档账号画像。",
        "operationId": "get_account_api_v1_accounts__url_name__get",
        "parameters": [
          {
            "name": "url_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Url Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Account Api V1 Accounts  Url Name  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{account_id}/timeline": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "Account Timeline",
        "description": "账号时间线（画像更新 + 踩饵命中记录）。",
        "operationId": "account_timeline_api_v1_accounts__account_id__timeline_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Account Timeline Api V1 Accounts  Account Id  Timeline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "alerts"
        ],
        "summary": "List Alerts",
        "description": "告警列表（级别/未读过滤）。含告警 payload（检测 id/分级），读取需 admin（P1-1）。",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(L3|L4|L5)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "unread_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unread Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Alerts Api V1 Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/read": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Mark Alert Read",
        "description": "标记已读。写操作需 admin（P1-1）。",
        "operationId": "mark_alert_read_api_v1_alerts__alert_id__read_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alert Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Mark Alert Read Api V1 Alerts  Alert Id  Read Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat-import/candidates": {
      "get": {
        "tags": [
          "chat-import"
        ],
        "summary": "Chat Candidates",
        "description": "探测微信/QQ 常见根路径（不扫全盘），返回 {candidates:{wechat,qq,custom}}。",
        "operationId": "chat_candidates_api_v1_chat_import_candidates_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat Candidates Api V1 Chat Import Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-import/scan": {
      "post": {
        "tags": [
          "chat-import"
        ],
        "summary": "Chat Scan",
        "description": "识别平台与消息库文件（含 SQLCipher 加密探测），返回 {result:{platform,db_files,encrypted}}。",
        "operationId": "chat_scan_api_v1_chat_import_scan_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatScanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat Scan Api V1 Chat Import Scan Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-import/import": {
      "post": {
        "tags": [
          "chat-import"
        ],
        "summary": "Chat Import",
        "description": "一键导入：解析 → 脱敏 → 落库 detections(chat_import) → 推理链分级。\n\n返回 {result:{imported,suspicious,preview}}；加密库 → error:'encrypted_db'，\n其它异常防御式回填 error 字段（服务层兜底，绝不 500）。",
        "operationId": "chat_import_api_v1_chat_import_import_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat Import Api V1 Chat Import Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-import/import-file": {
      "post": {
        "tags": [
          "chat-import"
        ],
        "summary": "Chat Import File",
        "description": "通用文本文件导入：parse_generic 预解析校验 → 复用 import_messages 全流程。\n\n返回 {result:{imported,suspicious,preview}}；无法解析 → 422 parse_failed。",
        "operationId": "chat_import_file_api_v1_chat_import_import_file_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatImportFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat Import File Api V1 Chat Import Import File Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat-import/imported": {
      "delete": {
        "tags": [
          "chat-import"
        ],
        "summary": "Chat Clear Imported",
        "description": "清空 chat_import 检测记录（写 gate_logs 审计），返回 {deleted:int}。\n\nconfirm!=true → 403 confirm_required；reason < 20 字 → 422 reason_too_short。",
        "operationId": "chat_clear_imported_api_v1_chat_import_imported_delete",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatClearRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chat Clear Imported Api V1 Chat Import Imported Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCheckRequest": {
        "properties": {
          "url_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Url Name",
            "description": "知乎 url_token/昵称"
          },
          "signals": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signals",
            "description": "公开信号（白名单过滤）：registered_at/activity_level/content_consistency/reported_count/trap_hit_ids 等"
          }
        },
        "type": "object",
        "required": [
          "url_name"
        ],
        "title": "AccountCheckRequest"
      },
...
      "ChatClearRequest": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "title": "ChatClearRequest",
        "description": "清空导入入参（必填）：{confirm:bool, reason:str}。"
      },
      "ChatImportFileRequest": {
        "properties": {
          "file_path": {
            "type": "string",
            "minLength": 1,
            "title": "File Path"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform"
          }
        },
        "type": "object",
        "required": [
          "file_path"
        ],
        "title": "ChatImportFileRequest",
        "description": "通用文本文件导入入参（必填）：{file_path, platform?}。"
      },
      "ChatImportRequest": {
        "properties": {
          "path": {
            "type": "string",
            "minLength": 1,
            "title": "Path"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 2000
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "ChatImportRequest",
        "description": "导入入参（必填）：{path, limit?=2000}。"
      },
      "ChatScanRequest": {
        "properties": {
          "path": {
            "type": "string",
            "minLength": 1,
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "ChatScanRequest",
        "description": "扫描入参（必填）：{path} —— 微信/QQ 目录或 TXT/CSV/JSON 文件路径。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}